Update Chat entity
Update Chat entity:
 - add Domain try/catch
 - fix Chat SignalR (leave/create chat connection)
 - fix Chat UI Issues

+++ src/API/CombatAnalysis.Hubs/Hubs/GroupChatMessagesHub.cs
@@ -65,7 +65,7 @@ public class GroupChatMessagesHub : Hub
                     Username = username,
                     Time = DateTimeOffset.UtcNow
                 },
-                State = (int)ChatMessageActionState.Created,
+                State = ChatMessageActionState.Created,
                 When = DateTimeOffset.UtcNow,
                 RefreshToken = Context.GetHttpContext()?.Request.Cookies[nameof(AuthenticationCookie.RefreshToken)] ?? string.Empty,
                 AccessToken = Context.GetHttpContext()?.Request.Cookies[nameof(AuthenticationCookie.AccessToken)] ?? string.Empty
